Windows installation details and fixes
Added a ToC.
Removed Node and NPM from requirements table.
Updated the instructions for the installation of Node and NPM. Simplified instructions, removed the instruction to check within the config file. This is unnecessary and installation can be verified with nvm -v. Added installation of cURL. Included additional steps to give user a better understanding of nvm.
Added details in the Windows section explaining the need for WSL and links to useful guides.
Updated vscode config file.
Increased all header levels by one.